refactor: move cli to cmd package
make help
make build with cross compilation
make tests with coverage
