+++ .gitignore
@@ -1,10 +1,16 @@
 .idea
 .vscode
-cli/venom/bin
-venom
+cmd/venom/bin
 tests/pprof*
 tests/.DS_Store
 .DS_Store
 vendor/
 dist/
 venom.log
+*.so
+foo.yml
+dist
+results
+bin.test
+profile.coverprofile
+tests.log
